+++ projeto-aluguel-de-carros/ms-cars/src/main/java/com/bandeira/aluguelcarros/dto/UpdateCarDTO.java
@@ -2,11 +2,13 @@ package com.bandeira.aluguelcarros.dto;
 
 import com.bandeira.aluguelcarros.model.Category;
 
+import java.math.BigDecimal;
+
 public record UpdateCarDTO(
 
         String name,
 
-        Double dailyPrice,
+        BigDecimal dailyPrice,
 
         Boolean available,
